Sync little refactor
